papermind/backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
sentence-transformers
scikit-learn
numpy
python-dotenv
pydantic
aiofiles
supabase
torch
transformers>=4.21.0
tokenizers
